Ai/brush-buddy/requirements.txt
annotated-types==0.6.0
greenlet==3.0.3
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.0
SQLAlchemy==2.0.25
typing==3.7.4.3
typing_extensions==4.9.0
annotated-types==0.6.0
anyio==4.2.0
black==23.12.1
boto3==1.34.26
botocore==1.34.26
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
isort==5.13.2
jmespath==1.0.1
multiprocessing-generator==0.3
mypy-extensions==1.0.0
numpy==1.26.3
openai==1.9.0
opencv-python==4.9.0.80
packaging==23.2
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.1.0
pydantic==2.5.3
pydantic_core==2.14.6
pypipo==0.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.0
scipy==1.12.0
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
tomli==2.0.1
tqdm==4.66.1
typing==3.7.4.3
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.27.0
watchfiles==0.21.0
websockets==12.0
